Refactor abstract base class so that more tests can reuse the base functionality
