+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/cache/BrokerEntryCacheRollingRestartTest.java
@@ -19,38 +19,28 @@
 package org.apache.pulsar.broker.cache;
 
 import static org.testng.Assert.assertTrue;
-import com.google.common.collect.Sets;
-import io.netty.channel.Channel;
 import io.netty.channel.EventLoopGroup;
 import java.io.Closeable;
 import java.io.File;
-import java.net.InetSocketAddress;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.LongSupplier;
 import java.util.stream.Collectors;
 import lombok.Cleanup;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.bookkeeper.client.PulsarMockBookKeeper;
-import org.apache.bookkeeper.client.api.LedgerEntries;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.BooleanUtils;
 import org.apache.pulsar.broker.BrokerTestUtil;
-import org.apache.pulsar.broker.MultiBrokerTestZKBaseTest;
-import org.apache.pulsar.broker.ServiceConfiguration;
 import org.apache.pulsar.broker.qos.AsyncTokenBucket;
-import org.apache.pulsar.broker.testcontext.PulsarTestContext;
 import org.apache.pulsar.client.api.Consumer;
 import org.apache.pulsar.client.api.InjectedClientCnxClientBuilder;
 import org.apache.pulsar.client.api.Message;
@@ -59,14 +49,9 @@ import org.apache.pulsar.client.api.Producer;
 import org.apache.pulsar.client.api.PulsarClient;
 import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.api.Schema;
-import org.apache.pulsar.client.api.SizeUnit;
 import org.apache.pulsar.client.api.SubscriptionInitialPosition;
 import org.apache.pulsar.client.api.SubscriptionType;
 import org.apache.pulsar.client.impl.PulsarClientImpl;
-import org.apache.pulsar.common.policies.data.TenantInfoImpl;
-import org.apache.pulsar.common.util.netty.ChannelFutures;
-import org.testng.SkipException;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Factory;
 import org.testng.annotations.Test;
 
@@ -76,205 +61,43 @@ import org.testng.annotations.Test;
  * The current intent is to show how the PIP-430 caching results in better cache hit rates in rolling restarts
  * when there's an active producer producing messages to a topic.
  * To avoid OOME, run this test on the command line (after enabling the test case) with this command:
- * ENABLE_MANUAL_TEST=true NETTY_LEAK_DETECTION=off mvn -pl pulsar-broker test -Dtest=BrokerEntryCacheMultiBrokerTest
+ * ENABLE_MANUAL_TEST=true NETTY_LEAK_DETECTION=off mvn -pl pulsar-broker test -Dtest=BrokerEntryCacheRollingRestartTest
  */
-@Test(groups = "broker-api")
 @Slf4j
-public class BrokerEntryCacheMultiBrokerTest extends MultiBrokerTestZKBaseTest {
-    public enum BrokerEntryCacheType {
-        PIP430("PIP430", BrokerEntryCacheMultiBrokerTest::configurePIP430),
-        PIP430Disabled("PIP430disabled", BrokerEntryCacheMultiBrokerTest::disablePIP430),
-        CacheEvictionByMarkDeletedPosition("cacheEvictionByMarkDeletedPosition",
-                BrokerEntryCacheMultiBrokerTest::configureCacheEvictionByMarkDeletedPosition),
-        PR12258Caching("PR12258", BrokerEntryCacheMultiBrokerTest::configurePR12258Caching);
-
-        private final String description;
-        private final java.util.function.Consumer<ServiceConfiguration> configurer;
-
-        BrokerEntryCacheType(String description, java.util.function.Consumer<ServiceConfiguration> configurer) {
-            this.description = description;
-            this.configurer = configurer;
-        }
-
-        public void configure(ServiceConfiguration config) {
-            configurer.accept(config);
-        }
-
-        public String getDescription() {
-            return description;
-        }
-    }
-
-    /**
-     * Configures PIP430 caching explicitly, although it's enabled by default.
-     * @param conf ServiceConfiguration instance to be configured
-     */
-    private static void configurePIP430(ServiceConfiguration conf) {
-        // use cache eviction by expected read count, this is the default behavior so it's not necessary to set it,
-        // but it makes the test more explicit
-        conf.setCacheEvictionByExpectedReadCount(true);
-        // LRU cache eviction behavior is enabled by default, but we set it explicitly
-        conf.setManagedLedgerCacheEvictionExtendTTLOfRecentlyAccessed(true);
-    }
-
-    /**
-     * Disables PIP430 caching features.
-     * @param conf ServiceConfiguration instance to be configured
-     */
-    private static void disablePIP430(ServiceConfiguration conf) {
-        conf.setCacheEvictionByExpectedReadCount(false);
-        conf.setManagedLedgerCacheEvictionExtendTTLOfRecentlyAccessed(false);
-    }
-
-    /**
-     * Configures ServiceConfiguration for cache eviction based on the slowest markDeletedPosition.
-     * This method disables cache eviction by expected read count and enables eviction by markDeletedPosition.
-     * Related to https://github.com/apache/pulsar/pull/14985 - "Evicting cache data by the slowest markDeletedPosition"
-     *
-     * @param conf ServiceConfiguration instance to be configured
-     */
-    private static void configureCacheEvictionByMarkDeletedPosition(ServiceConfiguration conf) {
-        disablePIP430(conf);
-        conf.setCacheEvictionByMarkDeletedPosition(true);
-    }
-
-    /**
-     * Configures ServiceConfiguration with settings to test PR12258 behavior for caching to drain backlog consumers.
-     * This method sets configurations to enable caching for cursors with backlogged messages.
-     * To make PR12258 effective, there's an additional change made in the broker codebase to
-     * activate the cursor when a consumer connects, instead of waiting for the scheduled task to activate it.
-     * Check org.apache.pulsar.broker.service.persistent.PersistentSubscription#addConsumerInternal method
-     * for the change.
-     * @param conf ServiceConfiguration instance to be modified
-     */
-    private static void configurePR12258Caching(ServiceConfiguration conf) {
-        disablePIP430(conf);
-        conf.setManagedLedgerMinimumBacklogCursorsForCaching(1);
-        conf.setManagedLedgerMinimumBacklogEntriesForCaching(1);
-        conf.setManagedLedgerMaxBacklogBetweenCursorsForCaching(Integer.MAX_VALUE);
-        conf.setManagedLedgerCursorBackloggedThreshold(Long.MAX_VALUE);
-    }
-
-    AtomicInteger bkReadCount = new AtomicInteger(0);
-    AtomicInteger bkReadEntryCount = new AtomicInteger(0);
-    volatile MessageId lastPublishedMessageId;
-    volatile boolean restarting;
-    private static final DateTimeFormatter CSV_DATETIME_FORMATTER =
-            DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSS");
-
+public class BrokerEntryCacheRollingRestartTest extends AbstractBrokerEntryCacheMultiBrokerTest {
     // comment out next line to run a single test run for the default cache type
     @Factory
     public static Object[] createTestInstances() {
-        return Arrays.stream(BrokerEntryCacheType.values()).map(BrokerEntryCacheMultiBrokerTest::new)
+        return Arrays.stream(BrokerEntryCacheType.values()).map(BrokerEntryCacheRollingRestartTest::new)
                 .toArray();
     }
 
-    private final BrokerEntryCacheType cacheType;
+    volatile MessageId lastPublishedMessageId;
+    volatile boolean restarting;
 
-    public BrokerEntryCacheMultiBrokerTest() {
+    public BrokerEntryCacheRollingRestartTest() {
         this(BrokerEntryCacheType.PIP430);
     }
 
-    public BrokerEntryCacheMultiBrokerTest(BrokerEntryCacheType cacheType) {
-        this.cacheType = cacheType;
+    public BrokerEntryCacheRollingRestartTest(BrokerEntryCacheType cacheType) {
+        super(cacheType);
     }
 
     @Override
-    protected void beforeSetup() {
-        if (!BooleanUtils.toBoolean(System.getenv("ENABLE_MANUAL_TEST")) && !isRunningInIntelliJ()) {
-            throw new SkipException("This test requires setting ENABLE_MANUAL_TEST=true environment variable.");
-        }
-    }
-
-    static boolean isRunningInIntelliJ() {
-        // Check for IntelliJ-specific system properties
-        return System.getProperty("idea.test.cyclic.buffer.size") != null
-                || System.getProperty("java.class.path", "").contains("idea_rt.jar");
-    }
-
-    @Override
-    protected void additionalSetup() throws Exception {
-        super.additionalSetup();
-        admin.tenants().createTenant("my-property",
-                new TenantInfoImpl(Sets.newHashSet("appid1", "appid2"), Sets.newHashSet("test")));
-        admin.namespaces().createNamespace("my-property/my-ns");
-        bkReadCount.set(0);
-        bkReadEntryCount.set(0);
-        PulsarMockBookKeeper mockBookKeeper = pulsarTestContext.getMockBookKeeper();
-        mockBookKeeper.setReadHandleInterceptor(
-                (long ledgerId, long firstEntry, long lastEntry, LedgerEntries entries) -> {
-                    bkReadCount.incrementAndGet();
-                    int numberOfEntries = (int) (lastEntry - firstEntry + 1);
-                    bkReadEntryCount.addAndGet(numberOfEntries);
-                    log.info(
-                            "BK read for ledgerId {}, firstEntry {}, lastEntry {}, numberOfEntries {}, "
-                            + "last msgid {}, restarting={}",
-                            ledgerId, firstEntry, lastEntry, numberOfEntries, lastPublishedMessageId, restarting);
-                    return mockBookKeeper.getJfrReadHandleInterceptor()
-                            .interceptReadAsync(ledgerId, firstEntry, lastEntry, entries);
-                });
-        // disable delays in bookkeeper writes and reads since they can skew the test results
-        mockBookKeeper.setDefaultAddEntryDelayMillis(0L);
-        mockBookKeeper.setDefaultReadEntriesDelayMillis(0L);
-    }
-
-    @Override
-    protected void customizeMainPulsarTestContextBuilder(PulsarTestContext.Builder pulsarTestContextBuilder) {
-        super.customizeMainPulsarTestContextBuilder(pulsarTestContextBuilder);
-        pulsarTestContextBuilder.spyNoneByDefault();
-    }
-
-    @Override
-    protected int numberOfAdditionalBrokers() {
-        // test with 2 brokers in the cluster
-        return 1;
-    }
-
-    @Override
-    protected boolean useDynamicBrokerPorts() {
-        return false;
-    }
-
-    @BeforeMethod(alwaysRun = true)
     public void beforeMethod() {
-        bkReadCount.set(0);
-        bkReadEntryCount.set(0);
+        super.beforeMethod();
         lastPublishedMessageId = null;
         restarting = false;
     }
 
     @Override
-    protected void doInitConf() throws Exception {
-        super.doInitConf();
-        configureDefaults(conf);
+    protected void logBKRead(long ledgerId, long firstEntry, long lastEntry, int numberOfEntries) {
+        log.info(
+                "BK read for ledgerId {}, firstEntry {}, lastEntry {}, numberOfEntries {}, "
+                        + "last msgid {}, restarting={}",
+                ledgerId, firstEntry, lastEntry, numberOfEntries, lastPublishedMessageId, restarting);
     }
 
-    private void configureDefaults(ServiceConfiguration conf) {
-        // configure ledger rollover to avoid OOM in test runs and slowdowns due to GC happening frequently when
-        // available heap memory is low
-        conf.setManagedLedgerMinLedgerRolloverTimeMinutes(0);
-        conf.setManagedLedgerMaxLedgerRolloverTimeMinutes(1);
-        conf.setManagedLedgerMaxEntriesPerLedger(100000);
-
-        // configure the cache type
-        cacheType.configure(conf);
-    }
-
-    @Override
-    protected ServiceConfiguration createConfForAdditionalBroker(int additionalBrokerIndex) {
-        ServiceConfiguration conf = super.createConfForAdditionalBroker(additionalBrokerIndex);
-        configureDefaults(conf);
-        return conf;
-    }
-
-    String serviceUrlForFixedPorts() {
-        return "pulsar://127.0.0.1:" + mainBrokerPort + "," + additionalBrokerPorts
-                .stream()
-                .map(port -> "127.0.0.1:" + port)
-                .collect(Collectors.joining(",")) + "/";
-    }
-
-    // change enabled to true to run the test
     @Test(invocationCount = 5)
     public void testTailingReadsRollingRestart() throws Exception {
         // this description is showed in result CSV files
@@ -520,17 +343,26 @@ public class BrokerEntryCacheMultiBrokerTest extends MultiBrokerTestZKBaseTest {
         assertTrue(consumersLatch.await(2, TimeUnit.SECONDS),
                 "All consumers should have been completed");
 
-        double cacheHitPercentage = 100.0d * (numberOfMessagesConsumed.get() - bkReadEntryCount.get())
-                / numberOfMessagesConsumed.get();
-        double cacheMissPercentage = 100.0d - cacheHitPercentage;
+        logAndRecordResultsToCsv(testConfigDescriptionInResult, numberOfMessagesProducedFuture.getNow(0L),
+                numberOfMessagesConsumed.get(), numConsumers, actualNumberOfRestarts.get());
+    }
 
-        Object[] msgArgs = {testConfigDescriptionInResult, numberOfMessagesProducedFuture.get(30, TimeUnit.SECONDS),
-                numberOfMessagesConsumed.get(), numConsumers, bkReadCount.get(), bkReadEntryCount.get(),
+    private void logAndRecordResultsToCsv(String testConfigDescriptionInResult, long numberOfMessagesProduced,
+                                          int numberOfMessagesConsumed, int numConsumers, int actualNumberOfRestarts)
+            throws InterruptedException, ExecutionException, TimeoutException, IOException {
+        int readEntryCount = bkReadEntryCount.get();
+        double cacheHitPercentage = 100.0d * (numberOfMessagesConsumed - readEntryCount)
+                / numberOfMessagesConsumed;
+        double cacheMissPercentage = 100.0d - cacheHitPercentage;
+        int readCount = bkReadCount.get();
+
+        Object[] msgArgs = {testConfigDescriptionInResult, numberOfMessagesProduced,
+                numberOfMessagesConsumed, numConsumers, readCount, readEntryCount,
                 String.format("%.2f", cacheHitPercentage), String.format("%.2f", cacheMissPercentage),
-                actualNumberOfRestarts.get()};
+                actualNumberOfRestarts};
         log.info("[{}] Produced {} and Consumed {} messages (across {} consumers with unique subscriptions) in total. "
-                        + "Number of BK reads {} with {} entries. Cache hits {}%, Cache misses {}%, Number of "
-                        + "restarts {}", msgArgs);
+                + "Number of BK reads {} with {} entries. Cache hits {}%, Cache misses {}%, Number of "
+                + "restarts {}", msgArgs);
 
         // record results in csv files for later processing with DuckDB, for example:
         // { cat pulsar-broker/target/rolling_restarts_result_header.txt; \
@@ -544,34 +376,8 @@ public class BrokerEntryCacheMultiBrokerTest extends MultiBrokerTestZKBaseTest {
         File resultFile = new File("target/rolling_restarts_result_" + ts + ".csv");
         List<String> csvColumns = new ArrayList<>();
         Arrays.stream(msgArgs).map(Object::toString).forEach(csvColumns::add);
-        csvColumns.add(
-                CSV_DATETIME_FORMATTER.format(ZonedDateTime.ofInstant(Instant.ofEpochMilli(ts), ZoneOffset.UTC)));
+        csvColumns.add(formatTimestampForCsv(ts));
         String resultMessage = csvColumns.stream().collect(Collectors.joining("\t")) + "\n";
         FileUtils.write(resultFile, resultMessage, StandardCharsets.UTF_8);
-    }
-
-    private PulsarClientImpl createPulsarClient(EventLoopGroup eventLoopGroup, LongSupplier connectionDelaySupplier)
-            throws Exception {
-        return InjectedClientCnxClientBuilder.builder()
-                .eventLoopGroup(eventLoopGroup)
-                .clientBuilder(
-                        PulsarClient.builder()
-                                .serviceUrl(serviceUrlForFixedPorts())
-                                .statsInterval(0, TimeUnit.SECONDS)
-                                .memoryLimit(32, SizeUnit.MEGA_BYTES)
-                )
-                .connectToAddressFutureCustomizer((Channel channel, InetSocketAddress inetSocketAddress) -> {
-                    // Introduce a delay to simulate network latency in connecting to the broker.
-                    long delayMillis = connectionDelaySupplier.getAsLong();
-                    if (delayMillis > 0) {
-                        return CompletableFuture.supplyAsync(() -> null,
-                                        CompletableFuture.delayedExecutor(delayMillis, TimeUnit.MILLISECONDS))
-                                .thenCompose(
-                                        __ -> ChannelFutures.toCompletableFuture(channel.connect(inetSocketAddress)));
-                    } else {
-                        return ChannelFutures.toCompletableFuture(channel.connect(inetSocketAddress));
-                    }
-                })
-                .build();
     }
 }
